+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Add `CallableTree::Node::Internal#seek?` to check whether the node's strategy is `seek` or not.
+- Add `CallableTree::Node::Internal#broadcast?` to check whether the node's strategy is `broadcast` or not.
 
 ## [0.3.0] - 2021-12-27